# ACL: Increase performance by selecting on indexed column